#6 - Database connection bug fixes

+++ physio2go/app/src/main/java/com/cm_project/physio2go/MainActivity.java
@@ -1,12 +1,16 @@
 package com.cm_project.physio2go;
 
+import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.net.ConnectivityManager;
+import android.net.NetworkInfo;
 import android.os.Build;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
+import android.view.View;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.RequiresApi;
@@ -20,6 +24,7 @@ import com.cm_project.physio2go.classes.Patient;
 import com.cm_project.physio2go.classes.Plan;
 import com.cm_project.physio2go.databaseDrivers.LocalDatabase;
 import com.cm_project.physio2go.databaseDrivers.ServerDatabaseDriver;
+import com.google.android.material.snackbar.Snackbar;
 
 import java.util.ArrayList;
 
@@ -31,6 +36,29 @@ public class MainActivity extends AppCompatActivity {
     public static final String LOGIN_NAME = "loggedin_username";
     final int REQ_LOGIN = 1;
     LocalDatabase local;
+
+    protected static boolean isNetworkAvilable(Context context) {
+        boolean isNetworkAvilable = false;
+        ConnectivityManager manager = (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
+        NetworkInfo info = manager.getActiveNetworkInfo();
+        if (info != null && info.isAvailable() && info.isConnected()) {
+            isNetworkAvilable = true;
+        }
+        return isNetworkAvilable;
+    }
+
+    public static void showNoInternetSnackbar(View v, String message) {
+
+        Snackbar snackBar = Snackbar
+                .make(v, message, Snackbar.LENGTH_INDEFINITE)
+                .setAction("Dismiss", new View.OnClickListener() {
+                    @Override
+                    public void onClick(View v) {
+
+                    }
+                });
+        snackBar.show();
+    }
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -47,47 +75,9 @@ public class MainActivity extends AppCompatActivity {
             Intent intent = new Intent(this, LoginActivity.class);
             startActivityForResult(intent, REQ_LOGIN);
         } else {
-            updateLocalDatabase(this.loggedInUsername);
-            inflateMainActivity();
+            boolean serverUpdated = updateLocalDatabase(this.loggedInUsername);
+            inflateMainActivity(serverUpdated);
         }
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent intent) {
-        super.onActivityResult(requestCode, resultCode, intent);
-
-        switch (requestCode){
-            case REQ_LOGIN:
-                if (resultCode == RESULT_OK) { // If login successfull
-                    if (intent!= null) {
-                        this.loggedInUsername = intent.getStringExtra("username");
-                        saveLoggedInUsername(this.loggedInUsername);
-                        updateLocalDatabase(this.loggedInUsername);
-                        inflateMainActivity();
-                    }
-                }
-                break;
-            default:
-                break;
-        }
-    }
-
-    private void inflateMainActivity() {
-        setContentView(R.layout.activity_main);
-
-        // Define a new toolbar for this activity
-        Toolbar toolbar = findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-
-        // Get plans of user from local DB
-        ArrayList<Plan> plansOfUser = local.getPlansOfUser();
-
-        // Instanciate list fragment
-        Fragment plansFragment = PlansListFragment.newInstance(plansOfUser);
-        FragmentTransaction ft = this.getSupportFragmentManager().beginTransaction();
-        ft.replace(R.id.fragment_list_placeholder, plansFragment);
-        //ft.addToBackStack(PlansListFragment.PLAN_LIST_FRAGMENT_TAG);
-        ft.commit();
     }
 
     @Override
@@ -154,31 +144,82 @@ public class MainActivity extends AppCompatActivity {
         editor.apply();
     }
 
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, Intent intent) {
+        super.onActivityResult(requestCode, resultCode, intent);
+
+        switch (requestCode) {
+            case REQ_LOGIN:
+                if (resultCode == RESULT_OK) { // If login successfull
+                    if (intent != null) {
+                        this.loggedInUsername = intent.getStringExtra("username");
+                        saveLoggedInUsername(this.loggedInUsername);
+                        boolean serverUpdated = updateLocalDatabase(this.loggedInUsername);
+                        inflateMainActivity(serverUpdated);
+                    }
+                }
+                break;
+            default:
+                break;
+        }
+    }
+
+    @RequiresApi(api = Build.VERSION_CODES.N)
+    public void getLocalDatabase() {
+        ArrayList<Plan> plansFromLocalDatabase = local.getPlansOfUser();
+    }
+
+    private void inflateMainActivity(boolean isServerUpdated) {
+        setContentView(R.layout.activity_main);
+
+        // Define a new toolbar for this activity
+        Toolbar toolbar = findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+
+        // Get plans of user from local DB
+        ArrayList<Plan> plansOfUser = local.getPlansOfUser();
+
+        // Instanciate list fragment
+        Fragment plansFragment = PlansListFragment.newInstance(plansOfUser);
+        FragmentTransaction ft = this.getSupportFragmentManager().beginTransaction();
+        ft.replace(R.id.fragment_list_placeholder, plansFragment);
+        //ft.addToBackStack(PlansListFragment.PLAN_LIST_FRAGMENT_TAG);
+        ft.commit();
+
+        // If server was not update, show Snackbar
+        if (!isServerUpdated) {
+            showNoInternetSnackbar(findViewById(R.id.ActivityLayout), "Could not synchronize with the Server.");
+        }
+    }
+
     /**
      * Extracts information from the server database and updates the local database
      * @param username
      */
-    public void updateLocalDatabase(String username){
-        ServerDatabaseDriver server = new ServerDatabaseDriver();
-
-        // Retrieve data from server
-        ArrayList<Plan> plansFromServer = server.getPlansOfUser(username);
-        Patient patient = server.getPatientDetails(username);
-
+    public boolean updateLocalDatabase(String username) {
         // Persist on local database
         local = new LocalDatabase(getApplicationContext());
 
-        // Delete files from localDatabase
-        local.delete();
+        boolean isNetAvailable = isNetworkAvilable(this);
 
-        // Updates plans and exercises
-        local.updatePlans(plansFromServer);
-        local.updatePatientDetails(patient);
-    }
+        if (isNetAvailable) {
+            // Retrieve data from server
+            ServerDatabaseDriver server = new ServerDatabaseDriver();
 
-    @RequiresApi(api = Build.VERSION_CODES.N)
-    public void getLocalDatabase(){
-        ArrayList<Plan> plansFromLocalDatabase = local.getPlansOfUser();
+            ArrayList<Plan> plansFromServer = server.getPlansOfUser(username);
+            Patient patient = server.getPatientDetails(username);
+
+            // Delete files from localDatabase
+            local.delete();
+
+            // Updates plans and exercises
+            local.updatePlans(plansFromServer);
+            local.updatePatientDetails(patient);
+
+            return true;
+        } else {
+            return false;
+        }
     }
 }
